refactor: PublicCircuitPublicInputs cleanup
