+++ yarn-project/circuits.js/src/structs/public_circuit_public_inputs.ts
@@ -2,7 +2,7 @@ import { makeTuple } from '@aztec/foundation/array';
 import { AztecAddress } from '@aztec/foundation/aztec-address';
 import { isArrayEmpty } from '@aztec/foundation/collection';
 import { Fr } from '@aztec/foundation/fields';
-import { BufferReader, Tuple, serializeToBuffer } from '@aztec/foundation/serialize';
+import { Tuple, serializeToBuffer } from '@aztec/foundation/serialize';
 import { FieldsOf } from '@aztec/foundation/types';
 
 import {
@@ -15,134 +15,7 @@ import {
   RETURN_VALUES_LENGTH,
 } from '../constants.gen.js';
 import { CallContext } from './call_context.js';
-import { Header, SideEffect, SideEffectLinkedToNoteHash } from './index.js';
-
-/**
- * Contract storage read operation on a specific contract.
- *
- * Note: Similar to `PublicDataRead` but it's from the POV of contract storage so we are not working with public data
- * tree leaf index but storage slot index.
- */
-export class ContractStorageRead {
-  constructor(
-    /**
-     * Storage slot we are reading from.
-     */
-    public readonly storageSlot: Fr,
-    /**
-     * Value read from the storage slot.
-     */
-    public readonly currentValue: Fr,
-    /**
-     * Optional side effect counter tracking position of this event in tx execution.
-     */
-    public readonly sideEffectCounter?: number,
-  ) {}
-
-  static from(args: {
-    /**
-     * Storage slot we are reading from.
-     */
-    storageSlot: Fr;
-    /**
-     * Value read from the storage slot.
-     */
-    currentValue: Fr;
-    /**
-     * Optional side effect counter tracking position of this event in tx execution.
-     */
-    sideEffectCounter?: number;
-  }) {
-    return new ContractStorageRead(args.storageSlot, args.currentValue, args.sideEffectCounter);
-  }
-
-  toBuffer() {
-    return serializeToBuffer(this.storageSlot, this.currentValue);
-  }
-
-  static fromBuffer(buffer: Buffer | BufferReader) {
-    const reader = BufferReader.asReader(buffer);
-    return new ContractStorageRead(Fr.fromBuffer(reader), Fr.fromBuffer(reader));
-  }
-
-  static empty() {
-    return new ContractStorageRead(Fr.ZERO, Fr.ZERO);
-  }
-
-  isEmpty() {
-    return this.storageSlot.isZero() && this.currentValue.isZero();
-  }
-
-  toFriendlyJSON() {
-    return `Slot=${this.storageSlot.toFriendlyJSON()}: ${this.currentValue.toFriendlyJSON()}`;
-  }
-}
-
-/**
- * Contract storage update request for a slot on a specific contract.
- *
- * Note: Similar to `PublicDataUpdateRequest` but it's from the POV of contract storage so we are not working with
- * public data tree leaf index but storage slot index.
- */
-export class ContractStorageUpdateRequest {
-  constructor(
-    /**
-     * Storage slot we are updating.
-     */
-    public readonly storageSlot: Fr,
-    /**
-     * Old value of the storage slot.
-     */
-    public readonly oldValue: Fr,
-    /**
-     * New value of the storage slot.
-     */
-    public readonly newValue: Fr,
-    /**
-     * Optional side effect counter tracking position of this event in tx execution.
-     */
-    public readonly sideEffectCounter?: number,
-  ) {}
-
-  toBuffer() {
-    return serializeToBuffer(this.storageSlot, this.oldValue, this.newValue);
-  }
-
-  static fromBuffer(buffer: Buffer | BufferReader) {
-    const reader = BufferReader.asReader(buffer);
-    return new ContractStorageUpdateRequest(Fr.fromBuffer(reader), Fr.fromBuffer(reader), Fr.fromBuffer(reader));
-  }
-
-  /**
-   * Create PublicCallRequest from a fields dictionary.
-   * @param fields - The dictionary.
-   * @returns A PublicCallRequest object.
-   */
-  static from(fields: FieldsOf<ContractStorageUpdateRequest>): ContractStorageUpdateRequest {
-    return new ContractStorageUpdateRequest(...ContractStorageUpdateRequest.getFields(fields));
-  }
-
-  /**
-   * Serialize into a field array. Low-level utility.
-   * @param fields - Object with fields.
-   * @returns The array.
-   */
-  static getFields(fields: FieldsOf<ContractStorageUpdateRequest>) {
-    return [fields.storageSlot, fields.oldValue, fields.newValue, fields.sideEffectCounter] as const;
-  }
-
-  static empty() {
-    return new ContractStorageUpdateRequest(Fr.ZERO, Fr.ZERO, Fr.ZERO);
-  }
-
-  isEmpty() {
-    return this.storageSlot.isZero() && this.oldValue.isZero() && this.newValue.isZero();
-  }
-
-  toFriendlyJSON() {
-    return `Slot=${this.storageSlot.toFriendlyJSON()}: ${this.oldValue.toFriendlyJSON()} => ${this.newValue.toFriendlyJSON()}`;
-  }
-}
+import { ContractStorageRead, ContractStorageUpdateRequest, Header, SideEffect, SideEffectLinkedToNoteHash } from './index.js';
 
 /**
  * Public inputs to a public circuit.
